pyproject.toml
[project]
name = "alpaca-py"
version = "0.0.0"
description = "The Official Python SDK for Alpaca APIs"
authors = [
    { name = "Rahul Chowdhury", email = "rahul.chowdhury@alpaca.markets" }
]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.8.0"
dependencies = [
    "requests>=2.30.0",
    "pydantic>=2.0.3",
    "pandas>=1.5.3",
    "msgpack>=1.0.3",
    "websockets>=10.4",
    "sseclient-py>=1.7.2",
    "python-dotenv>=1.0.1",
]

[project.urls]
Repository = "https://github.com/alpacahq/alpaca-py"
Documentation = "https://alpaca.markets/docs/python-sdk/"

[project.optional-dependencies]
dev = [
    "pytest>=7.1",
    "pytest-asyncio>=0.23.7",
    "requests-mock>=1.9.3",
    "black>=24.3.0",
    "isort>=5.10.1",
    "pre-commit>=2.17.0",
    "furo>=2022.2.14",
    "Sphinx>=5.0.1",
    "sphinx-copybutton>=0.5.0",
    "enum-tools>=0.9.0",
    "sphinx-toolbox>=3.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["alpaca"]

[dependency-groups]
dev = [
    "flake8>=5.0.4",
    "flake8-bugbear>=23.3.12",
    "flake8-comprehensions>=3.15.0",
    "flake8-import-order>=0.19.2",
]
